rename test files with spec
