+++ test/api/fulfillment.spec.ts
@@ -1,7 +1,7 @@
-import "../support/setup";
+import "../utils/setup";
 import { assert } from "chai";
 import { suite, test } from "mocha";
-import { mainApi } from "../constants";
+import { mainApi } from "../utils/constants";
 
 suite("Generating fulfillment data", () => {
   test(`Generate fulfillment data for listing`, async () => {
