+++ test/integration/getAsset.spec.ts
