+++ test/integration/getCollection.spec.ts
