+++ test/sdk/orders.spec.ts
@@ -4,8 +4,8 @@ import { suite, test } from "mocha";
 import { OpenSeaSDK } from "../../src/index";
 import { Network, OrderJSON } from "../../src/types";
 import { orderFromJSON } from "../../src/utils/utils";
-import { ALCHEMY_API_KEY, MAINNET_API_KEY } from "../constants";
 import ordersJSONFixture = require("../fixtures/orders.json");
+import { ALCHEMY_API_KEY, MAINNET_API_KEY } from "../utils/constants";
 
 const ordersJSON = ordersJSONFixture as unknown as OrderJSON[];
