+++ test/utils/constants.ts
@@ -1,5 +1,5 @@
-import { OpenSeaAPI } from "../src/api";
-import { Network } from "../src/types";
+import { OpenSeaAPI } from "../../src/api";
+import { Network } from "../../src/types";
 
 export const MAINNET_API_KEY = process.env.API_KEY;
 export const WALLET_PRIV_KEY = process.env.WALLET_PRIV_KEY;
